run gyr space change

+++ feature_extraction_ABIDE.py
@@ -502,26 +502,25 @@ def download_interTVA_from_frioul(
     -----
 
     """
-    # To be removed, tfMRI data
-    # if not os.path.exists("{}/{}/splitted/".format(intermediary_folder, subject)):
-    #     os.makedirs("{}/{}/splitted/".format(intermediary_folder, subject))
-
-    # for number in range(1, 178):
-    #     cmd = "scp mahaut.m@frioul.int.univ-amu.fr:/envau/work/banco/data/mri/InterTVA/my_intertva/surf/data/{0}/glm/vol/u{0}_task-localizer_model-singletrial_denoised/beta_{1:0>4}.nii.gz {2}/{0}/splitted/{0}__Res{1:0>4}.nii.gz".format(
-    #         subject, number, intermediary_folder
-    #     )
-    #     os.system(cmd)
-
     # Adding freesurfer directory
     for key in data_list["freesurfer"]:
         path = "{}/{}/{}".format(destination_folder, subject, key)
         if not os.path.exists(path):
             os.makedirs(path)
         for file in data_list["freesurfer"][key]:
-            cmd = "scp mahaut.m@frioul.int.univ-amu.fr:/envau/work/banco/data/mri/InterTVA/my_intertva/surf/data/{1}/fs/{1}/{2}/{3} {0}/{1}/{2}/{3}".format(
-                destination_folder, subject, key, file
-            )
-            os.system(cmd)
+            if not os.path.exists(
+                "{0}/{1}/{2}/{3}".format(destination_folder, subject, key, file)
+            ):
+                cmd = "scp mahaut.m@frioul.int.univ-amu.fr:/envau/work/banco/data/mri/InterTVA/my_intertva/surf/data/{1}/fs/{1}/{2}/{3} {0}/{1}/{2}/{3}".format(
+                    destination_folder, subject, key, file
+                )
+                os.system(cmd)
+            else:
+                print(
+                    "{} allready exists. Remove file if you wish to download again.".format(
+                        file
+                    )
+                )
     print("Downloaded all of {}'s required files".format(subject))
